Laget felles komponent for knapperad.
Fjernet registrert-tidspunkt fra kort.
