+++ src/app/components/observation/observation/observation.component.css
@@ -97,20 +97,3 @@ ion-chip.chip-disabled {
 .registration {
   margin-top: 12px;
 }
-
-.buttons {
-  display: flex;
-  /* Knapperaden har ion-color-light klasse som endrer --ion-background-color */
-  background-color: var(--ion-background-color);
-}
-
-.reg-id {
-  margin-left: auto;
-}
-
-/*
- * Hvis det bare er én knapp i knapperaden (kopier-regid-knapp)
- */
-.buttons:has(> :last-child:nth-child(1)) {
-  flex-direction: row-reverse;
-}
