--- src/app/components/observation/registration-view-config.model.ts
@@ -1,38 +1,0 @@
-import { Type } from '@angular/core';
-import { RegistrationTid } from 'src/app/modules/common-registration/models/registration-tid.enum';
-import { RegistrationViewModel } from 'src/app/modules/common-regobs-api';
-
-export interface RegistrationViewConfig {
-  /**
-   * ID til skjemaet som skal vises
-   */
-  tid: RegistrationTid;
-
-  /**
-   * Hvilken komponent som skal vises
-   */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  component: Type<any>;
-
-  /**
-   * Funksjon som tar i mot en registrering og sier fra om komponenten bør rendres eller ikke.
-   * (Det er i alle fall det den brukes til i dag. Kunne kanskje endra navn til "shouldRenderComponent", eller noe.).
-   *
-   * @param registration Registrering / observasjon
-   */
-  isEmpty: (registration: RegistrationViewModel) => boolean;
-
-  /**
-   * Funksjon som tar i mot en registrering og gjør klar inputs til komponenten.
-   *
-   * @param registration Registrering / observasjon
-   * @returns Et objekt med alle inputs til komponenten.
-   */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  getInputs: (registration: RegistrationViewModel) => any;
-
-  /**
-   * Skal returnere en språknøkkel eller allerede oversatt tekst til tittel
-   */
-  getHeader: (registration: RegistrationViewModel) => string;
-}
